S3Cost_Optimisation_Forecast/requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
langchain-community
langgraph-groq

